Re #2407 Move exceptions out of Stack.Types.Config
