+++ src/Stack/Lock.hs
@@ -24,7 +24,7 @@ import           Path.Extended ( addExtension )
 import           Path.IO ( doesFileExist )
 import           Stack.Prelude
 import           Stack.SourceMap ( snapToDepPackage )
-import           Stack.Types.Config ( ConfigPrettyException (..) )
+import           Stack.Types.Config.Exception ( ConfigPrettyException (..) )
 import           Stack.Types.LockFileBehavior ( LockFileBehavior (..) )
 import           Stack.Types.Runner ( HasRunner, lockFileBehaviorL, rslInLogL )
 import           Stack.Types.SourceMap ( DepPackage, SMWanted )
